feat!: v0.9.0 breaking changes (#242, #243, #244)
- Extract FailureClassifier trait to tower-resilience-core (#242)
- Remove deprecated CircuitBreakerRequestLayer (#243)
- Standardize event_type() returns to snake_case (#244)

Closes #242, closes #243, closes #244
