+++ crates/tower-resilience-circuitbreaker/src/layer.rs
@@ -229,67 +229,3 @@ impl<S, F> Layer<S> for CircuitBreakerLayer<FnClassifier<F>> {
         CircuitBreaker::new(service, Arc::clone(&self.config))
     }
 }
-
-// Keep the old API for backwards compatibility, but mark as deprecated
-/// Request-typed circuit breaker layer for backwards compatibility.
-///
-/// This type is **deprecated**. Use `CircuitBreakerLayer` directly instead,
-/// which now implements `Layer<S>` without requiring `.for_request()`.
-#[deprecated(
-    since = "0.7.0",
-    note = "CircuitBreakerLayer now implements Layer<S> directly. Use CircuitBreakerLayer::builder().build() instead."
-)]
-#[derive(Clone)]
-pub struct CircuitBreakerRequestLayer<Req, C> {
-    config: Arc<CircuitBreakerConfig<C>>,
-    _phantom: std::marker::PhantomData<fn() -> Req>,
-}
-
-#[allow(deprecated)]
-impl<S, Req, C> Layer<S> for CircuitBreakerRequestLayer<Req, C>
-where
-    C: Clone,
-{
-    type Service = CircuitBreaker<S, C>;
-
-    fn layer(&self, service: S) -> Self::Service {
-        CircuitBreaker::new(service, Arc::clone(&self.config))
-    }
-}
-
-impl<C> CircuitBreakerLayer<C>
-where
-    C: Clone,
-{
-    /// Converts this layer into a request-typed layer.
-    ///
-    /// **Deprecated**: This method is no longer needed. `CircuitBreakerLayer` now
-    /// implements `Layer<S>` directly.
-    ///
-    /// # Migration
-    ///
-    /// Before:
-    /// ```rust,ignore
-    /// ServiceBuilder::new()
-    ///     .layer(layer.for_request::<MyRequest>())
-    ///     .service(my_service)
-    /// ```
-    ///
-    /// After:
-    /// ```rust,ignore
-    /// ServiceBuilder::new()
-    ///     .layer(layer)
-    ///     .service(my_service)
-    /// ```
-    #[deprecated(
-        since = "0.7.0",
-        note = "CircuitBreakerLayer now implements Layer<S> directly. Remove .for_request() call."
-    )]
-    #[allow(deprecated)]
-    pub fn for_request<Req>(&self) -> CircuitBreakerRequestLayer<Req, C> {
-        CircuitBreakerRequestLayer {
-            config: Arc::clone(&self.config),
-            _phantom: std::marker::PhantomData,
-        }
-    }
-}
